refactor(agent): extract PTT key formatting to shared utility
